GH-467 Move permission request form base and common logic to shared module
